% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{pls}
\alias{pls}
\title{Partial Least Squares (PLS)}
\usage{
pls(X, Y, num.components)
}
\arguments{
\item{X}{Matrix of descriptors (n x p)}

\item{Y}{Matrix of response variables (n x q)}

\item{num.components}{Number of PLS components (variates) to compute}
}
\value{
Named list containing weights, scores, loadings, rotations and coefficients
}
\description{
This function implements the generic PLS algorithm
}
